requirements.txt
python-osc
websocket-client